+++ crates/icrate/CHANGELOG.md
@@ -15,6 +15,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
   View the release notes to learn more details:
   - [15.1](https://developer.apple.com/documentation/xcode-release-notes/xcode-15_1-release-notes)
+* Marked `NSView::isFlipped`, `NSView::convertRect_toView`,
+  `NSWindow::convertRectToScreen` and `NSWindow::convertPointFromScreen` as
+  safe.
 
 ### Deprecated
 * Deprecated `MainThreadMarker::run_on_main`, use the new free-standing function